Refactored Provider to its own folder
